requirements.txt
fastapi
uvicorn
python-multipart
streamlit
requests
python-dotenv
pillow
torch
transformers
langchain-groq
ultralytics